Rename filename according to conventions.
